project-mortgage-pipeline/requirements.txt
# Kafka
kafka-python-ng==2.2.2

# Data processing
pyspark==3.5.1
delta-spark==3.2.0
pandas==2.2.2
pyarrow==16.1.0

# API & HTTP
requests==2.32.3
python-dotenv==1.0.1

# Data quality
great-expectations==0.18.19

# Observability
prometheus-client==0.20.0

# Testing
pytest==8.2.2
pytest-mock==3.14.0

# Code quality
ruff==0.4.9